# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  description: Logistics API provides services to find and manage entities for logistics
  title: Logistics API
  version: "1.0"
  contact:
    email: vladislavprokopenko1@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
...
  /orders:
    get:
      tags:
        - Order
      summary: Find all Orders
      operationId: ordersFindAll
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders were not found
        500:
          description: Internal Server Error

    post:
      tags:
        - Order
      summary: Create a new Order
      operationId: orderCreate
      requestBody:
        description: Order object that needs to be added to order list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        201:
          description: Order was created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /orders/{orderId}:
    patch:
      tags:
        - Order
      summary: Update Order
      operationId: orderUpdate
      parameters:
        - name: orderId
          in: path
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      requestBody:
        description: Order object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrder'
      responses:
        200:
          description: Order was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error

    delete:
      tags:
        - Order
      summary: Delete Order
      operationId: orderDelete
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        204:
          description: Order was deleted successfully
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error

    get:
      tags:
        - Order
      summary: Find Order by id
      operationId: orderFindById
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        200:
          description: Order was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error
  /orders/byCustomer/{orderCustomerId}:
    get:
      tags:
        - Order
      summary: Find Orders by orderCustomerId
      operationId: orderFindByOrderCustomerId
      parameters:
        - name: orderCustomerId
          in: path
          description: "Customer id of Order"
          required: true
          schema:
            type: integer
            format: int32
            example: 5001
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderCustomerId was not found
        500:
          description: Internal Server Error
  /orders/category/{category}:
    get:
      tags:
        - Order
      summary: Find Orders by category
      operationId: orderFindByCategory
      parameters:
        - name: category
          in: path
          description: "Order category"
          required: true
          schema:
            type: string
            example: Papers
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with defined category were not found
        500:
          description: Internal Server Error
  /orders/status/{status}:
    get:
      tags:
        - Order
      summary: Find Orders by status
      operationId: orderFindByStatus
      parameters:
        - name: status
          in: path
          description: "Order status. Possible statuses: NEW, EXPECT_DRIVERS_CONFIRMATION, CONFIRMED, DECLINED_BY_DRIVERS, ON_ROAD, COMPLETED"
          required: true
          schema:
            type: string
            example: NEW
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with defined category were not found
        500:
          description: Internal Server Error
  /orders/truckNumber/{assignedTruckNumber}:
    get:
      tags:
        - Order
      summary: Find Order by assignedTruckNumber
      operationId: orderFindByAssignedTruckNumber
      parameters:
        - name: assignedTruckNumber
          in: path
          description: "Assigned Truck number for Order"
          required: true
          schema:
            type: string
            example: ABC1234
      responses:
        200:
          description: Order was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderCustomerId was not found
        500:
          description: Internal Server Error
  /orders/earlier/startDateTime/{startDateTime}:
    get:
      tags:
        - Order
      summary: Find Orders with startDateTime earlier or equal to defined
      operationId: orderFindByEarlierThanStartDateTime
      parameters:
        - name: startDateTime
          in: path
          description: "Start date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with start date time earlier or equal to defined were not found
        500:
          description: Internal Server Error
  /orders/later/startDateTime/{startDateTime}:
    get:
      tags:
        - Order
      summary: Find Orders with startDateTime later or equal to defined
      operationId: orderFindByLaterThanStartDateTime
      parameters:
        - name: startDateTime
          in: path
          description: "Start date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with start date time later or equal to defined were not found
        500:
          description: Internal Server Error
  /orders/earlier/limitDateTime/{limitDateTime}:
    get:
      tags:
        - Order
      summary: Find Orders with limitDateTime earlier or equal to defined
      operationId: orderFindByEarlierThanLimitDateTime
      parameters:
        - name: limitDateTime
          in: path
          description: "Limit date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with limit date time earlier or equal to defined were not found
        500:
          description: Internal Server Error
  /orders/later/limitDateTime/{limitDateTime}:
    get:
      tags:
        - Order
      summary: Find Orders with limitDateTime later or equal to defined
      operationId: orderFindByLaterThanLimitDateTime
      parameters:
        - name: limitDateTime
          in: path
          description: "Limit date time"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Orders were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Orders with limit date time later or equal to defined were not found
        500:
          description: Internal Server Error
  /orders/status/{orderId}:
    patch:
      tags:
        - Order
      summary: Update Order status
      operationId: orderUpdateStatus
      parameters:
        - name: orderId
          in: path
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      requestBody:
        description: Order object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatus'
      responses:
        200:
          description: Order status was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error
  /orders/driverComment/{orderId}:
    patch:
      tags:
        - Order
      summary: Update driver comment for Order
      operationId: orderUpdateDriverComment
      parameters:
        - name: orderId
          in: path
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      requestBody:
        description: Order object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderDriverComment'
      responses:
        200:
          description: Driver comment for Order was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error
  /orders/setStartDateTime/{orderId}:
    patch:
      tags:
        - Order
      summary: Set current date-time as Order start date-time
      operationId: orderUpdateStartDateTime
      parameters:
        - name: orderId
          in: path
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        200:
          description: Order start date-time was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error
  /orders/assignedTruckNumber/{orderId}:
    patch:
      tags:
        - Order
      summary: Update assigned truck number for Order
      operationId: orderUpdateAssignedTruckNumber
      parameters:
        - name: orderId
          in: path
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      requestBody:
        description: Order object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderAssignedTruckNumber'
      responses:
        200:
          description: Assigned truck number for Order was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This orderId was not found
        500:
          description: Internal Server Error
...
components:
  schemas:
...
    CreateOrder:
      description: Object representing Order creating
      type: object
      properties:
        orderCustomerId:
          type: integer
          format: int32
          example: 5001
        category:
          type: string
          example: Papers
        weight:
          description: Total order weight, tons
          type: number
          format: double
          example: 14.2
        limitDateTime:
          type: string
          format: date-time
    UpdateOrder:
      description: Object representing Order updating
      type: object
      properties:
        orderCustomerId:
          type: integer
          format: int32
          example: 5001
        category:
          type: string
          example: Papers
        weight:
          description: Total order weight, tons
          type: number
          format: double
          example: 14.2
        status:
          type: string
          enum: [NEW,EXPECT_DRIVERS_CONFIRMATION,CONFIRMED,DECLINED_BY_DRIVERS,ON_ROAD,COMPLETED]
          default: NEW
        limitDateTime:
          type: string
          format: date-time
    UpdateOrderDriverComment:
      description: Object representing updating driver comment for Order
      type: object
      properties:
        driverComment:
          description: Driver comment
          type: string
          example: "I'm not going to drive with Sam"
    UpdateOrderAssignedTruckNumber:
      description: Object representing updating assigned truck number for Order
      type: object
      properties:
        assignedTruckNumber:
          description: Assigned truck number
          type: string
          example: "ABC1234"
    UpdateOrderStatus:
      description: Object representing Order updating
      type: object
      properties:
        status:
          type: string
          enum: [NEW,EXPECT_DRIVERS_CONFIRMATION,CONFIRMED,DECLINED_BY_DRIVERS,ON_ROAD,COMPLETED]
          default: NEW
    Order:
      description: Object representing Order
      type: object
      properties:
        orderId:
          type: integer
          format: int32
          example: 12001
        orderCustomerId:
          type: integer
          format: int32
          example: 5001
        category:
          type: string
          example: Papers
        weight:
          description: Total order weight, tons
          type: number
          format: double
          example: 14.2
        status:
          type: string
          enum: [NEW,EXPECT_DRIVERS_CONFIRMATION,CONFIRMED,DECLINED_BY_DRIVERS,ON_ROAD,COMPLETED]
          default: NEW
        startDateTime:
          type: string
          format: date-time
        limitDateTime:
          type: string
          format: date-time
        assignedTruckNumber:
          type: string
          example: ABC1234
        driverComment:
          type: string
          example: I'm not going to drive with Sam
...